requirements.txt
pymupdf==1.24.10
pdfminer.six==20240706